+++ resources/views/layouts/navigation.blade.php
@@ -24,7 +24,6 @@
                     <x-nav.link url="{{ route('customer.dashboard', $customer) }}" :name="__('Daschboard')"><x-svg.home class="h-6 w-6" /> </x-nav.link>
                     <x-nav.link url="{{ route('customer.search') }}" :name="__('Kundensuche')"><x-svg.search class="h-6 w-6" /> </x-nav.link>
                     <x-nav.link url="{{ route('search.global') }}" :name="__('Globale Suche')"><x-svg.db class="h-6 w-6" /> </x-nav.link>
-                    <x-nav.link url="{{ route('search.utm') }}" :name="__('UTM Suche')" target="_blank"><x-svg.software.securepoint class="h-6 w-6" /> </x-nav.link>
                     <x-nav.link url="{{ route('search.remote') }}" :name="__('Rustdesk Suche')" target="_blank"><x-svg.software.rustdesk class="h-6 w-6" /> </x-nav.link>
 
                 </div>